fix: defer hart_delete to checkpoint, fix block_diff overflow, eliminate double root
Three fixes and two improvements:

1. hart_delete deferred to checkpoint blocks only — state_finalize_block
   no longer deletes dead accounts from acct_index. This prevents the
   duplicate-account bug where hart_delete + ensure_account creates a
   new slot, orphaning the original's storage/code. Manifested as root
   mismatch at ~2.7M blocks (Spurious Dragon EIP-161 pruning).

2. block_diff group_count/slot_count widened from uint16_t to uint32_t —
   the Shanghai cleanup blocks produce 258K+ account diffs per block,
   overflowing uint16_t and causing a buffer overflow crash.

3. collect_block_diff uses blk_dirty_cursor to only process current
   block's entries, not accumulated entries across checkpoint window.
   Prevents bloated 48M-entry diffs and 137GB history files.

4. Eliminate double root computation — sync layer now uses br.state_root
   from block_execute instead of recomputing via evm_state_compute_mpt_root.

5. Compaction gated on !skip_root_hash so it only runs after root
   computation. snapshot-every aligned to checkpoint boundaries.

+++ executor/include/block_diff.h
@@ -52,14 +52,14 @@ typedef struct {
     hash_t      code_hash;   /* new code_hash (valid if field_mask & FIELD_CODE_HASH) */
     hash_t      old_code_hash; /* previous code_hash (for undo) */
     slot_diff_t *slots;      /* heap-allocated array */
-    uint16_t    slot_count;
+    uint32_t    slot_count;
 } addr_diff_t;
 
 /** Per-block diff: array of address groups. */
 typedef struct block_diff_t {
     uint64_t     block_number;
     addr_diff_t *groups;        /* heap-allocated array */
-    uint16_t     group_count;
+    uint32_t     group_count;
 } block_diff_t;
 
 /** Free a block_diff_t's heap allocations (not the struct itself). */
